Fix empty stream list after first login
When a user logs in for the first time, the stream list was not populated.
That's because fillStreams() was being called with userId=0.

The user ID is set on the Settings object just prior to fillStreams() is called, but Settings.commit() wasn't called, so the newly set user ID was not actually committed until after fillStreams() is finished, resulting in fillStreams() failing silently due to userId=0.